Add files via upload
Problem Statement
String Analysis
You are given a string that represents an email address: "My e-mail: eMail_Address321@anymail.com". Your task is to analyze the composition of the characters in the string and determine the percentage of uppercase letters, lowercase letters, digits, and other characters such as symbols (#$., etc).
To accomplish this, you need to break down the string and calculate the percentage for each category. The results are as follows:
Uppercase letters: 7.5% Lowercase letters: 65% Digits: 7.5% Other characters (symbols): 20%